Refactor(typechecker): reorganize Pass 1 into per-construct bricks
Split the `primitive` module along AST constructs, mirroring the
`flowlog-build` catalog module's layout, so each file is one language
construct and the dependencies flow upward:

  ty      the LitKind lattice (foundation)
  expr    infer + pin an expression (Arithmetic/Factor)
  atom    bind/check/pin an atom's arguments
  fact    check/pin declared fact tuples
  compare check/pin a body comparison
  rule    the composer: check/pin a rule and its head

`mod.rs` now holds no logic — module declarations, the shared private
`Bindings` alias, and the one whole-program config gate. `check_program`
calls `rule::check_and_pin_rules` and `fact::check_and_pin_facts`
directly, so there is no driver/wrapper layer.

Naming is unified to `<verb>_<node>` (infer/check/pin), with
`check_and_pin_<node>` where a function does both; Pass 2's per-rule
walk becomes `check_and_lower_rule`. Visibility is a three-tier scheme:
`pub(crate)` for the entries `lib.rs` calls, `pub(super)` for helpers
shared within `primitive`, private for locals and `Bindings`.

Module docs trimmed to one-liners; the crate-level reject/allow lists
are now comprehensive (every user-facing `TypeCheckError` variant,
grouped by pass, with examples).

Behavior-preserving: pure code movement + renames, no logic change.
Verified by `cargo build`, `cargo clippy --all-targets -D warnings`, the
crate's 23 unit/integration tests, and the full fixture suite (121
programs x both the compiler and lib runners, byte-identical output).
